refactor: create useFetchLocations
used useSWR to fetch locations, cooldownbutton revalidate the data by using mutate funciton with keys
